Making up/down work better and tweaking <leader>ev
Now up/down will scroll window like <C-E>/<C-Y> while moving the cursor
along the way so that it stays in the same window column.

Also changing <leader>ev to not just open the vim rcs folder but also to
cd to it so things like Ag will actually search the rcs folder instead
of whatever folder we happened to be in when initially starting vim
(usually a project repo or some such).