Removing "command" bits. Copy/pasted lines from SO but don't need it
"command" is a built-in command that says to execute the following
binary. So if there were a binary "git" in the path but also a local
function or something that was also called "git", the only way to
forcibly execute the binary, instead of the function, would be to run
`command git`. But we don't have any functions or anything else to
conflict so... why bother.