Fixing bug with colors in prompt screwing up newlines in readline
Apparently when you do escape sequences in PROMPT (and probably other
things as well), you need to wrap them in %{...%} in order to signal to
readline that this isn't taking up space. Otherwise if you get to the
end of the line in a prompt, it starts overwriting your first line for
ab it before going down to the next line.