Turning off cursorline for quickfix window
Cursorline being on in quickfix window is obnoxious because I like
having the currently selected item in the quickfix window having a nice
heavy background color. If cursorline is on and the cursor is on the
currently active item the nice muted background is used instead of the
heavy QuickFixLine.